Added fps counter
Very basic (and most likely incorrect logic). Just to test the text quad for the now.

I have decided to leave the coordinate system for the now as it is a minor issue. However, the uniforms for text colour needs to be set and a new font is required.

Working on this has presented several issues on the meshdata class and I am looking into potential ways to split it up.