Split GLShader into an interface
Interface requires further work + testing

The aim is to try and get the required functionality that will work between vulkan and Opengl. I think this class should do the following:

Set up when to be used by the renderer
Handle setting of info required to be rendered
Handle linking/ creation of shader
Drawing of a mesh.

What needs to be tidied up:
Split functionality about using lighting information.