Implementing window surface for vulkan
Running the vulkan renderer will not break the game as I have removed OpenGL calls. The reason it fails is due to the game trying to create OpenGL shaders when glew has been removed.

The current thinking is to have the shaders get created via the renderer so that the vulkan implementation is seperate from OpenGL.

However, framebuffers/ textures & modelInfo will need to be moved. There are going to be many breaking changes from this point until the code base has been fully refactored.